refactor: MenuGroup 과 Menu 분리로 테스트 코드 패키지 분리
