+++ src/test/java/kitchenpos/Acceptance/AcceptanceTest.java
@@ -14,7 +14,9 @@ import org.springframework.test.context.ActiveProfiles;
 
 import java.util.Collections;
 
-import static kitchenpos.menu.MenuGenerator.*;
+import static kitchenpos.menu.MenuGenerator.메뉴_상품_생성_요청;
+import static kitchenpos.menu.MenuGenerator.메뉴_생성_API_호출;
+import static kitchenpos.menuGroup.MenuGroupGenerator.메뉴_그룹_생성_API_호출;
 import static kitchenpos.product.ProductGenerator.상품_생성_API_요청;
 import static kitchenpos.table.TableGenerator.주문_테이블_생성_요청;
 import static kitchenpos.table.TableGenerator.테이블_생성_API_호출;
