+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -20,7 +20,9 @@ import java.util.Collections;
 import java.util.List;
 
 import static kitchenpos.common.domain.PriceTest.가격_생성;
-import static kitchenpos.menu.MenuGenerator.*;
+import static kitchenpos.menu.MenuGenerator.메뉴_상품_생성_요청;
+import static kitchenpos.menu.MenuGenerator.메뉴_생성_요청;
+import static kitchenpos.menuGroup.MenuGroupGenerator.메뉴_그룹_생성;
 import static kitchenpos.product.ProductGenerator.상품_생성;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
